Refactore project folder structure
