+++ src/components/message/index.tsx
@@ -1,4 +1,4 @@
-import { IResponseMessageFromAI } from "@/services/types";
+import { IResponseMessageFromAI } from "@/api/types";
 import { UserCircle } from "lucide-react";
 import Image from "next/image";
 import { cn, formatDate } from "@/lib/utils";
